requirements.txt
fastapi
uvicorn
requests
sentence-transformers
faiss-cpu
pydantic
numpy